\name{scraEP-package}
\alias{scraEP-package}
\docType{package}

\title{
Scrape the Web with Extra Power.
}

\description{
A set of tools for scraping information from webpages and other XML contents.
}

\details{

\tabular{ll}{
Package: \tab scraEP\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2021-06-21\cr
License: \tab GPL (>=3)\cr
}

Function \code{xscrape} is a general tool to extract information from html pages into data frames using XPath or CSS queries.

Function \code{unaccent} removes accents from character strings.

Function \code{strcomp} compares the elements of two character vectors.
}

\author{
Julien Boelaert \email{jubo.stats@gmail.com}
}

\keyword{ package }
