\name{wiki}
\alias{wiki}
\docType{data}
\title{Wikipedia page for R.}

\description{Toy example to extract data using \code{xscrape}.}

\usage{
data(wiki)
}

\format{
Object \code{wiki} is a character vector.
}

\details{
Object \code{wiki} is a raw webpage, containing the source of the 
`R (programming language)` article on English wikipedia 
(<https://en.wikipedia.org/wiki/R_(programming_language)>, retrieved on 15/11/2017).
}

\author{
Julien Boelaert \email{jubo.stats@gmail.com}
}

\keyword{datasets}
