\name{scraEP-package}
\alias{scraEP-package}
\docType{package}

\title{
Scrape the careers of all members of European parliament.
}

\description{
Fetch all in-house career information of MEPs from the European parliament's website, and put them in a data frame.
}

\details{

\tabular{ll}{
Package: \tab scraEP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-11-15\cr
License: \tab GPL (>=3)\cr
}

Function \code{scraEP} downloads and extracts career information from the EP's website.

Function \code{xscrape} is a general tool to extract information from html pages into data frames using XPath queries.
}

\author{
Julien Boelaert \email{jubo.stats@gmail.com}
}

\keyword{ package }
